refactor(lib::podcast): move struct "NewEpisode" to "episode_db" as it is only used in db
